+++ src/main/java/com/splitspendings/groupexpensesbackend/dto/appuser/AppUserReceivedGroupInvitesDto.java
@@ -1,6 +1,6 @@
 package com.splitspendings.groupexpensesbackend.dto.appuser;
 
-import com.splitspendings.groupexpensesbackend.dto.groupinvite.GroupInviteDto;
+import com.splitspendings.groupexpensesbackend.dto.group.invite.GroupInviteDto;
 import lombok.Data;
 
 import java.util.List;
