+++ src/main/java/com/splitspendings/groupexpensesbackend/dto/group/invite/GroupInviteAcceptedDto.java
@@ -1,0 +1,10 @@
+package com.splitspendings.groupexpensesbackend.dto.group.invite;
+
+import com.splitspendings.groupexpensesbackend.dto.group.membership.GroupMembershipDto;
+import lombok.Data;
+
+@Data
+public class GroupInviteAcceptedDto {
+
+    private GroupMembershipDto groupMembership;
+}
