+++ src/main/java/com/splitspendings/groupexpensesbackend/dto/group/membership/GroupMembershipShortDto.java
@@ -1,14 +1,12 @@
-package com.splitspendings.groupexpensesbackend.dto.groupmembership;
+package com.splitspendings.groupexpensesbackend.dto.group.membership;
 
 import com.splitspendings.groupexpensesbackend.dto.appuser.AppUserDto;
 import com.splitspendings.groupexpensesbackend.dto.group.GroupInfoDto;
 import lombok.Data;
 
 @Data
-public class GroupMembershipWithIdShortDto {
+public class GroupMembershipShortDto {
 
-    private Long id;
-    private Boolean active;
     private GroupInfoDto group;
     private AppUserDto appUser;
 }
