+++ src/main/java/com/splitspendings/groupexpensesbackend/dto/group/membership/settings/GroupMembershipSettingsDto.java
@@ -1,17 +1,13 @@
-package com.splitspendings.groupexpensesbackend.dto.groupmembershipsettings;
+package com.splitspendings.groupexpensesbackend.dto.group.membership.settings;
 
 import com.splitspendings.groupexpensesbackend.model.enums.GroupTheme;
 import com.splitspendings.groupexpensesbackend.model.enums.NotificationOption;
 import lombok.Data;
 
-import javax.validation.constraints.NotNull;
-
 @Data
-public class UpdateGroupMembershipSettingsDto {
+public class GroupMembershipSettingsDto {
 
-    @NotNull
+    private Long id;
     private GroupTheme groupTheme;
-
-    @NotNull
     private NotificationOption notificationOption;
 }
