+++ src/main/java/com/splitspendings/groupexpensesbackend/dto/group/membership/settings/UpdateGroupMembershipSettingsDto.java
@@ -1,13 +1,17 @@
-package com.splitspendings.groupexpensesbackend.dto.groupmembershipsettings;
+package com.splitspendings.groupexpensesbackend.dto.group.membership.settings;
 
 import com.splitspendings.groupexpensesbackend.model.enums.GroupTheme;
 import com.splitspendings.groupexpensesbackend.model.enums.NotificationOption;
 import lombok.Data;
 
-@Data
-public class GroupMembershipSettingsDto {
+import javax.validation.constraints.NotNull;
 
-    private Long id;
+@Data
+public class UpdateGroupMembershipSettingsDto {
+
+    @NotNull
     private GroupTheme groupTheme;
+
+    @NotNull
     private NotificationOption notificationOption;
 }
