+++ src/main/java/com/splitspendings/groupexpensesbackend/dto/spending/comment/NewSpendingCommentDto.java
@@ -1,4 +1,4 @@
-package com.splitspendings.groupexpensesbackend.dto.spendingcomment;
+package com.splitspendings.groupexpensesbackend.dto.spending.comment;
 
 import com.splitspendings.groupexpensesbackend.model.SpendingComment;
 import com.splitspendings.groupexpensesbackend.util.TrimUtil;
@@ -8,7 +8,10 @@ import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
 @Data
-public class UpdateSpendingCommentDto {
+public class NewSpendingCommentDto {
+
+    @NotNull
+    private Long spendingId;
 
     @NotNull
     @Size(min = SpendingComment.MESSAGE_MIN_LENGTH, max = SpendingComment.MESSAGE_MAX_LENGTH)
