+++ src/main/java/com/splitspendings/groupexpensesbackend/dto/spending/comment/SpendingCommentDto.java
@@ -1,4 +1,4 @@
-package com.splitspendings.groupexpensesbackend.dto.spendingcomment;
+package com.splitspendings.groupexpensesbackend.dto.spending.comment;
 
 import com.splitspendings.groupexpensesbackend.dto.appuser.AppUserDto;
 import com.splitspendings.groupexpensesbackend.dto.spending.SpendingShortDto;
