+++ src/main/java/com/splitspendings/groupexpensesbackend/mapper/AppUserBalanceMapper.java
@@ -1,6 +1,6 @@
 package com.splitspendings.groupexpensesbackend.mapper;
 
-import com.splitspendings.groupexpensesbackend.dto.appuserbalance.AppUserBalanceDto;
+import com.splitspendings.groupexpensesbackend.dto.appuser.balance.AppUserBalanceDto;
 import com.splitspendings.groupexpensesbackend.model.UserBalance;
 import org.mapstruct.Mapper;
