+++ src/main/java/com/splitspendings/groupexpensesbackend/mapper/AppUserSettingsMapper.java
@@ -1,9 +1,9 @@
 package com.splitspendings.groupexpensesbackend.mapper;
 
-import com.splitspendings.groupexpensesbackend.dto.appusersettings.AppUserSettingsDto;
-import com.splitspendings.groupexpensesbackend.dto.appusersettings.AppUserSettingsWithIdDto;
-import com.splitspendings.groupexpensesbackend.dto.appusersettings.NewAppUserSettingsDto;
-import com.splitspendings.groupexpensesbackend.dto.appusersettings.UpdateAppUserSettingsDto;
+import com.splitspendings.groupexpensesbackend.dto.appuser.settings.AppUserSettingsDto;
+import com.splitspendings.groupexpensesbackend.dto.appuser.settings.AppUserSettingsWithIdDto;
+import com.splitspendings.groupexpensesbackend.dto.appuser.settings.NewAppUserSettingsDto;
+import com.splitspendings.groupexpensesbackend.dto.appuser.settings.UpdateAppUserSettingsDto;
 import com.splitspendings.groupexpensesbackend.model.AppUserSettings;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
