+++ src/main/java/com/splitspendings/groupexpensesbackend/mapper/GroupMembershipMapper.java
@@ -1,6 +1,6 @@
 package com.splitspendings.groupexpensesbackend.mapper;
 
-import com.splitspendings.groupexpensesbackend.dto.groupmembership.GroupMembershipDto;
+import com.splitspendings.groupexpensesbackend.dto.group.membership.GroupMembershipDto;
 import com.splitspendings.groupexpensesbackend.model.GroupMembership;
 import org.mapstruct.Mapper;
