+++ src/main/java/com/splitspendings/groupexpensesbackend/mapper/GroupMembershipSettingsMapper.java
@@ -1,7 +1,7 @@
 package com.splitspendings.groupexpensesbackend.mapper;
 
-import com.splitspendings.groupexpensesbackend.dto.groupmembershipsettings.GroupMembershipSettingsDto;
-import com.splitspendings.groupexpensesbackend.dto.groupmembershipsettings.UpdateGroupMembershipSettingsDto;
+import com.splitspendings.groupexpensesbackend.dto.group.membership.settings.GroupMembershipSettingsDto;
+import com.splitspendings.groupexpensesbackend.dto.group.membership.settings.UpdateGroupMembershipSettingsDto;
 import com.splitspendings.groupexpensesbackend.model.GroupMembershipSettings;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
