+++ src/main/java/com/splitspendings/groupexpensesbackend/service/AppUserBalanceService.java
@@ -1,6 +1,6 @@
 package com.splitspendings.groupexpensesbackend.service;
 
-import com.splitspendings.groupexpensesbackend.dto.appuserbalance.AppUserBalanceDto;
+import com.splitspendings.groupexpensesbackend.dto.appuser.balance.AppUserBalanceDto;
 import com.splitspendings.groupexpensesbackend.model.Group;
 import com.splitspendings.groupexpensesbackend.model.UserBalance;
