+++ src/main/java/com/splitspendings/groupexpensesbackend/service/AppUserService.java
@@ -7,8 +7,8 @@ import com.splitspendings.groupexpensesbackend.dto.appuser.AppUserGroupsDto;
 import com.splitspendings.groupexpensesbackend.dto.appuser.AppUserReceivedGroupInvitesDto;
 import com.splitspendings.groupexpensesbackend.dto.appuser.NewAppUserDto;
 import com.splitspendings.groupexpensesbackend.dto.appuser.UpdateLoginNameDto;
-import com.splitspendings.groupexpensesbackend.dto.appusersettings.AppUserSettingsWithIdDto;
-import com.splitspendings.groupexpensesbackend.dto.appusersettings.UpdateAppUserSettingsDto;
+import com.splitspendings.groupexpensesbackend.dto.appuser.settings.AppUserSettingsWithIdDto;
+import com.splitspendings.groupexpensesbackend.dto.appuser.settings.UpdateAppUserSettingsDto;
 import com.splitspendings.groupexpensesbackend.model.AppUser;
 import com.splitspendings.groupexpensesbackend.model.AppUserSettings;
