+++ src/main/java/com/splitspendings/groupexpensesbackend/service/GroupInviteService.java
@@ -1,8 +1,8 @@
 package com.splitspendings.groupexpensesbackend.service;
 
-import com.splitspendings.groupexpensesbackend.dto.groupinvite.GroupInviteAcceptedDto;
-import com.splitspendings.groupexpensesbackend.dto.groupinvite.GroupInviteDto;
-import com.splitspendings.groupexpensesbackend.dto.groupinvite.NewGroupInviteDto;
+import com.splitspendings.groupexpensesbackend.dto.group.invite.GroupInviteAcceptedDto;
+import com.splitspendings.groupexpensesbackend.dto.group.invite.GroupInviteDto;
+import com.splitspendings.groupexpensesbackend.dto.group.invite.NewGroupInviteDto;
 import com.splitspendings.groupexpensesbackend.model.GroupInvite;
 
 public interface GroupInviteService {
