+++ src/main/java/com/splitspendings/groupexpensesbackend/service/impl/AppUserServiceImpl.java
@@ -8,11 +8,11 @@ import com.splitspendings.groupexpensesbackend.dto.appuser.AppUserIdentityDto;
 import com.splitspendings.groupexpensesbackend.dto.appuser.AppUserReceivedGroupInvitesDto;
 import com.splitspendings.groupexpensesbackend.dto.appuser.NewAppUserDto;
 import com.splitspendings.groupexpensesbackend.dto.appuser.UpdateLoginNameDto;
-import com.splitspendings.groupexpensesbackend.dto.appusersettings.AppUserSettingsDto;
-import com.splitspendings.groupexpensesbackend.dto.appusersettings.AppUserSettingsWithIdDto;
-import com.splitspendings.groupexpensesbackend.dto.appusersettings.UpdateAppUserSettingsDto;
+import com.splitspendings.groupexpensesbackend.dto.appuser.settings.AppUserSettingsDto;
+import com.splitspendings.groupexpensesbackend.dto.appuser.settings.AppUserSettingsWithIdDto;
+import com.splitspendings.groupexpensesbackend.dto.appuser.settings.UpdateAppUserSettingsDto;
 import com.splitspendings.groupexpensesbackend.dto.group.GroupInfoDto;
-import com.splitspendings.groupexpensesbackend.dto.groupinvite.GroupInviteDto;
+import com.splitspendings.groupexpensesbackend.dto.group.invite.GroupInviteDto;
 import com.splitspendings.groupexpensesbackend.mapper.AppUserMapper;
 import com.splitspendings.groupexpensesbackend.mapper.AppUserSettingsMapper;
 import com.splitspendings.groupexpensesbackend.mapper.GroupInviteMapper;
